requirements.txt
0wdg9nbmpm==0.1
beautifulsoup4==4.6.0
bleach==2.0.0
certifi==2017.7.27.1
chardet==3.0.4
Django==1.11.4
django-crispy-forms==1.6.1
django-debug-toolbar==1.8
django-filer==1.3.0
django-htmlmin==0.10.0
django-mptt==0.8.7
django-polymorphic==1.3
django-tinymce==2.6.0
djangorestframework==3.6.3
easy-thumbnails==2.4.2
fake-factory==9999.9.9
foo==0.1
html5lib==0.999999999
httplib2==0.10.3
idna==2.6
Markdown==2.6.9
MySQL-python==1.2.5
olefile==0.44
path.py==10.3.1
Pillow==4.1.1
psycopg2==2.7.3
Pygments==2.2.0
pytz==2017.2
requests==2.18.4
six==1.10.0
sqlparse==0.2.3
test==1.0.1
Unidecode==0.4.21
urllib==1.21.1
urllib3==1.22
webencodings==0.5.1
xlrd==1.1.0
xlutils==2.0.0
xlwt==1.3.0
